Updated Scripts - Added BMS/Orion References
All scripts that exist in future versions of the game have had their NPC and Script updated, as well as a working version of the script in a different language to be ported to python. The remaining 21 npc's that weren't updated seem to be specific to this version and will be a hit or miss as far as what they do.