Updated NPC Scripts
All NPC scripts (with exceptions below) have been scripted. Previously broken or non-translated scripts have also been fixed and/or translated to English.

Job Changing scripts have yet to be ported from Orion (JS) => OrionAlpha (Python).
The JS version is documented inside the script and just needs to be converted over.

Hair Stylist scripts have yet to be implemented. These aren't implemented in Orion yet either, so Nexon's official scripts for these are contained in each of the stylist's script.

Refiner scripts are not implemented. As such, Nexon's official scripts for these are contained in each of the refiner's script. They just need to be converted to Python.