Updated ClientPackets (WIP)
This is an updated list of CP opcodes outside of a very small gap I haven't had time to dig in to the client for. Will update this gap once I've finished updating multiplayer packets.

Additionally, I've changed hexadecimal output to decimal seeing as our opcodes are in decimal format. Unfortunately, ExceptionLog was taken away from us as well :'(