Implemented Channel Changing
Full support for channel changing has been added now that it exists in the client. Without client edits, the only way to migrate is via /channel <channel>. However, with client edits, the channel change menu exists in the game files and can be modified to work in the options menu.