Updated README - Fixed Pet Action Rendering
After discontinuing pets, I ended up finding the template loader and realized that it was functional. So, I re-looked at the cause for my packets crashing only to pinpoint it to PrepareActionLayer causing a crash due to path resolving failure. After fixing the path, the action layers were successfully able to render, and thus pets are fully functional in the client.