Implemented Weapon Stickers
So when my original findings of this pointing to weapon stickers failed, it turns out that the equip encoded is not the sticker, but the unseen equip itself.. since I kept encoding the sticker, it never acknowledged it. After testing this theory out, I finally realized it really was stickers the whole time - so, now 3rd party cash weapon effects are working!